before adding database in the project
I have created routes for every functionality those ware needed for the project in future I will add more functionality. this server is not doing the many things, I think right now server is doing signup and sign in functionality and saving user in the current meory of the project.