after login testing.
When I was able to make user login the chat admin page where user can start the make his own chat. I realis that my mongo db client is not connecting to the database. in my last concern I was trying again to get the resion of login failed. and reasion is mongodb client which's connection is responsible for performing the opration in the database. is not being connecting to the mongodeb cluster. so login is failing may be ip blacklist ka chakkar ya fir mongodb atlus ki limit ho gai hogi. I will check it in next session of work.