this is After I create Documentation
Back end app is running. this time i need to check it online on internet need a VPS service now to test it online. I tried to run it on pc as server but it is not working My google and router is showing diffrent public ip address I tried to run both but not working now i am setting up my HP laptop to amke it server. so this commite for get the server files repo in laptop